Add an API for starting the playground
